server/pyproject.toml
[tool.poetry]
name = "synctabs"
version = "0.1.0"
description = ""
authors = ["asyncee"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.54.1"
jinja2 = "^2.11.2"
uvicorn = "^0.11.3"
bcrypt = "^3.1.7"
python-dotenv = "^0.12.0"
injector = "^0.18.3"
typer = "^0.1.1"
orjson = "^2.6.5"
aiofiles = "^0.5.0"

[tool.poetry.dev-dependencies]
pytest-cov = "^2.8.1"
mypy = "^0.770"
mypy-extensions = "^0.4.3"
flake8 = "^3.7.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 100
target-version = ['py38']
